fix(packaging): simplify Arch checks and path resolution
