+++ package.json
@@ -21,7 +21,6 @@
     "tauri:dev": "pnpm run build && concurrently -k -n SYSTEM,TAURI -c yellow,cyan \"pnpm dev:system-apps\" \"tauri dev\"",
     "tauri:build": "pnpm build && tauri build",
     "package:arch": "node scripts/package-arch.mjs",
-    "check:arch-package": "node scripts/check-arch-package.mjs",
     "lint": "eslint .",
     "preview": "vite preview",
     "prettier": "prettier --write .",
